Move op; remove ttnn::narrow path
