+++ tt-train/sources/ttml/ops/moe_ffn_swiglu_op.cpp
@@ -7,7 +7,7 @@
 #include <ttnn/operations/eltwise/binary/binary.hpp>
 #include <ttnn/operations/eltwise/unary/unary.hpp>
 
-#include "metal/ops/sparse_matmul/sparse_matmul.hpp"
+#include "ops/sparse_matmul_op.hpp"
 
 namespace ttml::ops {
 
@@ -18,15 +18,15 @@ ttnn::Tensor moe_ffn_swiglu_fw(
     const ttnn::Tensor& w_up,
     const ttnn::Tensor& w_down) {
     // Gate and up-projection into the intermediate dim.
-    auto gate_proj = ttml::metal::sparse_matmul(grouped, w_gate, offsets);
-    auto up_proj = ttml::metal::sparse_matmul(grouped, w_up, offsets);
+    auto gate_proj = sparse_matmul(grouped, w_gate, offsets);
+    auto up_proj = sparse_matmul(grouped, w_up, offsets);
 
     // SiLU(gate) * up. Pad rows are zero in both operands, so they stay zero.
     auto activated = ttnn::multiply(ttnn::silu(gate_proj), up_proj);
     gate_proj.deallocate();
     up_proj.deallocate();
 
-    auto out = ttml::metal::sparse_matmul(activated, w_down, offsets);
+    auto out = sparse_matmul(activated, w_down, offsets);
     activated.deallocate();
     return out;
 }
